# Change to fast.Sin() implementation with faster LUT/interpolation