--- a/go.mod
+++ b/go.mod
@@ -2,11 +2,11 @@ module github.com/synaptecltd/emulator
 
 go 1.18
 
-require github.com/stretchr/testify v1.7.0
+require github.com/stretchr/testify v1.8.1
 
 require (
-	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/teknico/sigourney v0.0.0-20160105013017-b52044632571
-	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )