Before with `math.Sin()`:
```
BenchmarkEmulator
BenchmarkEmulator-24
      24	   1992738 ns/op	     286 B/op	       0 allocs/op
PASS
ok  	github.com/synaptecltd/emulator	0.228s
```

After with `fast.Sin()`:
```
BenchmarkEmulator
BenchmarkEmulator-24
      24	    682038 ns/op	     286 B/op	       0 allocs/op
PASS
ok  	github.com/synaptecltd/emulator	0.197s
```

There doesn't appear to be any visible impact on the "smoothness" of generated waveforms.